moved signal functions to signals.c
