+++ src/executer.c
@@ -94,44 +94,6 @@ int	single_parent_process(t_msh *msh)
 	return (status);
 }
 
-void	sigquit_handler(int signal)
-{
-	//TODO need to 
-	if (signal == SIGQUIT)
-	{
-		g_exit_code = 131;
-		write(1, "\n", 1);
-		rl_on_new_line();
-		rl_replace_line("", 0);
-		rl_redisplay();
-	}
-}
-
-void	set_quit_action(void)
-{
-	struct sigaction sa_quit;
-
-	ft_bzero(&sa_quit, sizeof(sa_quit));
-	sa_quit.sa_handler = &sigquit_handler;
-	sigemptyset(&sa_quit.sa_mask);
-	sa_quit.sa_flags = SA_RESTART;
-	// TODO wrap up in a in if sigaction == -1 for error check
-	sigaction(SIGQUIT, &sa_quit, NULL);
-}
-
-struct sigaction	sigignore(int signal)
-{
-	struct sigaction	sa;
-	struct sigaction	sa_old;
-
-	ft_bzero(&sa, sizeof(sa));
-	sigemptyset(&sa.sa_mask);
-	sa.sa_handler = SIG_IGN;
-	sa.sa_flags = 0;
-	sigaction(signal, &sa, &sa_old);
-
-	return (sa_old);
-}
 
 void	child_process(t_msh *msh, int prev_pipe_read_fd, int *fd)
 {
@@ -232,7 +194,7 @@ int	process(t_msh *msh)
 		}
 		msh->command = msh->command->pipe_next;
 	}
-	status = wait_for_children(first_command);
+	wait_for_children(first_command);
 	//printf("status: %d\n", status);
 	sigaction(SIGINT, &sa_int_old, NULL);
 	sigaction(SIGQUIT, &sa_quit_old, NULL);
