+++ src/main.c
@@ -2,32 +2,6 @@
 
 volatile sig_atomic_t	g_exit_code;
 
-void	sigint_handler(int signal)
-{
-	//TODO need to 
-	if (signal == SIGINT)
-	{
-		g_exit_code = 130;
-		write(1, "\n", 1);
-		rl_on_new_line();
-		rl_replace_line("", 0);
-		rl_redisplay();
-	}
-}
-
-void	set_signal_action(void)
-{
-	struct sigaction sa_int;
-
-	ft_bzero(&sa_int, sizeof(sa_int));
-	sa_int.sa_handler = &sigint_handler;
-	sigemptyset(&sa_int.sa_mask);
-	sa_int.sa_flags = SA_RESTART;
-	// TODO wrap up in a in if sigaction == -1 for error check
-	sigaction(SIGINT, &sa_int, NULL);
-	signal(SIGQUIT, SIG_IGN);
-}
-
 int	main(int argc, char **argv, char **envp) {
 	char	*line;
 	t_msh	msh;
